Move enum ProverState to protocols layer
