+++ aries_vcx/src/handlers/mod.rs
@@ -2,13 +2,13 @@ use futures::future::BoxFuture;
 
 use crate::error::prelude::*;
 use crate::handlers::connection::connection::ConnectionState;
-use crate::handlers::proof_presentation::prover::ProverState;
 use crate::handlers::proof_presentation::verifier::VerifierState;
 use crate::messages::a2a::A2AMessage;
 use crate::protocols::connection::invitee::state_machine::InviteeState;
 use crate::protocols::connection::inviter::state_machine::InviterState;
 use crate::protocols::issuance::holder::state_machine::HolderState;
 use crate::protocols::issuance::issuer::state_machine::IssuerState;
+use crate::protocols::proof_presentation::prover::state_machine::ProverState;
 
 pub mod connection;
 pub mod issuance;
